Adjust namespaces, move JavaRandom
