+++ Minecraft Enchantment Cracker/Minecraft Enchantment Cracker.csproj
@@ -61,6 +61,7 @@
       <SubType>Designer</SubType>
     </ApplicationDefinition>
     <Compile Include="Tasks\ItemFinderTask.cs" />
+    <Compile Include="Tasks\JavaRandom.cs" />
     <Compile Include="Tasks\VersionTask.cs" />
     <Page Include="MainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
