+++ Minecraft Enchantment Cracker/Tasks/ItemFinderTask.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Minecraft_Enchantment_Cracker
+﻿namespace Minecraft_Enchantment_Cracker.Tasks
 {
     public class ItemFinderTask : IProgressiveTask {
